asmlinkage, x86: Remove leftover extern __visible
Remove the left over extern __visible Linus objected too.
They should be useless, as we only need the visible
on the actual definitions.

Mostly for x86, but one in export.h too.

Signed-off-by: Andi Kleen <ak@linux.intel.com>

Changes made for 3.19 update: A section was moved from hibernate_64
and others to sections.h, which was updated.  Looks like the original
patch missed hibernate_32.c which is now covered by sections.h too.

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>