meta: Add tinification/staging feature
Add staging patches, tinification patches which are in the process of
being getting readied for upstreaming to tinification but aren't quite
ready to be submitted or are 'ready' and submitted but haven't yet
been merged.

The tinification/staging branch is merged in via KERNEL_FEATURES as
features/staging.scc.  To have the branch merged and the features
available, add (for example):

  KERNEL_FEATURES_append_galileo += " features/tinification/staging.scc"

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>